CI Fix and Clean Up
Improved the CI for T3 to now ensure that the base OS used on GitHub has ample space for the installation and testing of T3. Additionally, we activate the clean command in RMG-Py to ensure that all files are not cythonised prior to cythonising. Finally, installation of Julia and the connection to Python and vice versa has been adjusted.