T3 Main Test Fix
T3 Main tests were failing due to changes in RMG-Py. This has been rectified. The first change is now we look for the numeric identifier of the species instead of the species name as the numeric identifier should never change and is unique. The second function change is that the chemical reaction string uses '<=>' instead of '='.